Add macros
Add macros to create color presets

The first macro creates presets of 16 colors for the first 6 device groups.

The second macro saves scenes on executors on page 5 - please enable "OFF ON OVERWRITTEN" in the global executor settings before running it.

The third macro creates the option to switch colors for high and low presets - (last two rows on page 5).

The remaining two macros create RGB presets for a greater variety of colors (test).